requirements.txt
streamlit
Pillow
ultralytics
numpy